Domains: Google Apps: move google-apps-user-item to its own file
