+++ client/my-sites/upgrades/domain-management/style.scss
@@ -997,11 +997,11 @@ ul.dns__list {
 	}
 }
 
-.google-apps-users-card__user-email {
+.google-apps-user-item__email {
 	float: left;
 }
 
-.google-apps-users-card__user-manage-link {
+.google-apps-user-item__manage-link {
 	float: right;
 	font-size: 13px;
 }
